SIMPLE still not working properly
Both the SIMPLE and SCGS are not stable and quickly diverge if:

- Too high under relaxation factors are chosen (>0.1)
- High Reynolds number (>100)
- Coarse mesh

For SCGS the problem can be spotted as a non-diagonal dominance of Vanka's matrix. For SIMPLE, an analysis of the residual between iteration, shows that the code starts converging (up to ~200 outer iterations) and the solution of the problem seems reasonable. Suddently some issue in the corner top-right of the the domain brings to a rapid divergence of the system.

At this point we believe that the issue is somewhere in the coefficients used for convection, and in particular about the computation of fluxes.